+++ test/test_compress_consistency_gtest.cpp
@@ -42,9 +42,9 @@ protected:
 
     void pump() {
         for (int i = 0; i < 10; i++) {
-            qwrt_tick(rt);
+            qwrt_tick(rt, 100);
             pal_mock_fire_all_timers(pal);
-            qwrt_tick(rt);
+            qwrt_tick(rt, 100);
         }
     }